Added files via upload
The purpose of this assignment was to prompt for a location, contact a web service and retrieve JSON for the web service and parse that data, and retrieve the first place_id from the JSON. A place ID is a textual identifier that uniquely identifies a place as within Google Maps. We used an API endpoint that has a static subset of the Google Data, found here: http://python-data.dr-chuck.net/geojson  And were asked to find the place identifier for "Indian School of Mines Dhanbad".