Added files via upload
The purpose of this assignment was to use urllib to read the HTML from the data files below, extract the href= vaues from the anchor tags, scan for a tag that is in a particular position relative to the first name in the list, follow that link and repeat the process a number of times and report the last name found.

Start at: http://python-data.dr-chuck.net/known_by_Loudon.html 
Find the link at position 18 (the first name is 1). Follow that link. Repeat this process 7 times. The answer is the last name that you retrieve, and will start with H.