Added files via upload
This assignment was from the "Python for Informatics" textbook by Charles Severance, p 138.  The program counts the distribution of the hour of the day for each of the messages in a txt file. After accumulating counts for each hour, the counts are printed, one per line and sorted by hour.