Added files via upload
This assignment comes from the textbook "Python for Informatics" by Charles Severance, on p 141.  The assignment for a program that will look for lines of the form: New Revision: 39772 in the text file mbox.txt and extract the number from each of the lines using a regular expression and the findall() method. Compute the average of the numbers and print out the average.